fix(controller): save before deleting old save file and abort (return) on failed IO on renaming operation

+++ src/com/lnw/acceeeeeppt/scene/menu/LoadGameController.java
@@ -63,9 +63,13 @@ public class LoadGameController {
                     break;
             }
             PlayerModel newPlayerModel = new PlayerModel(oldPlayerModel, newName);
+            int status = SaveManager.saveToDisk(newPlayerModel);
+            if (status != 0) {
+                loadGameView.showJOptionPaneWarning("Rename operation failed on saving, ABORTING.", "Rename Error");
+                return;
+            }
             SaveManager.deleteSave(oldPlayerModel);
             loadGameView.removeSelectedSaveSlotView();
-            SaveManager.saveToDisk(newPlayerModel);
             loadGameView.addSaveEntries(newPlayerModel.getSaveName(), newPlayerModel.getCurrLevel().toString(),
                     newPlayerModel.getCreatedDateTimeInstant().toString());
         } catch (ClassNotFoundException _) {
